+++ src/Cronofy/CronofyOAuthClient.cs
@@ -1,4 +1,4 @@
-﻿namespace Cronofy
+namespace Cronofy
 {
     using System;
     using System.Collections.Generic;
@@ -301,18 +301,7 @@
             }
         }
 
-        /// <summary>
-        /// Creates a smart invite for the given request.
-        /// </summary>
-        /// <param name="smartInviteRequest">
-        /// The details of the invite, must not be <code>null</code>.
-        /// </param>
-        /// <returns>
-        /// A smart invite for the given request.
-        /// </returns>
-        /// <exception cref="CronofyException">
-        /// Thrown if an error is encountered whilst making the request.
-        /// </exception>
+        /// <inheritdoc/>
         public SmartInvite CreateInvite(SmartInviteRequest smartInviteRequest)
         {
             var request = new HttpRequest
@@ -328,19 +317,12 @@
             return response.ToSmartInvite();
         }
 
-        /// <summary>
-        /// Cancels a smart invite for the given request.
-        /// </summary>
-        /// <param name="smartInviteId">The invite id to cancel.</param>
-        /// <param name="recipientEmail">The recipient for the cancellation.</param>
-        /// <returns>
-        /// A smart invite for the given request.
-        /// </returns>
-        /// <exception cref="CronofyException">
-        /// Thrown if an error is encountered whilst making the request.
-        /// </exception>
+        /// <inheritdoc/>
         public SmartInvite CancelInvite(string smartInviteId, string recipientEmail)
         {
+            Preconditions.NotEmpty("smartInviteId", smartInviteId);
+            Preconditions.NotEmpty("emailAddress", recipientEmail);
+
             var request = new HttpRequest
             {
                 Method = "POST",
@@ -356,23 +338,12 @@
             return response.ToSmartInvite();
         }
 
-        /// <summary>
-        /// Retreives detials of a smart invite.
-        /// </summary>
-        /// <param name="smartInviteId">
-        /// The invite id.
-        /// </param>
-        /// <param name="emailAddress">
-        /// The email address of the invitee.
-        /// </param>
-        /// <returns>
-        /// A smart invite for the given request.
-        /// </returns>
-        /// <exception cref="CronofyException">
-        /// Thrown if an error is encountered whilst making the request.
-        /// </exception>
+        /// <inheritdoc/>
         public SmartInvite GetSmartInvite(string smartInviteId, string emailAddress)
         {
+            Preconditions.NotEmpty("smartInviteId", smartInviteId);
+            Preconditions.NotEmpty("emailAddress", emailAddress);
+
             var request = new HttpRequest
             {
                 Method = "GET",
@@ -387,6 +358,43 @@
             request.AddOAuthAuthorization(this.clientSecret);
 
             var response = this.HttpClient.GetJsonResponse<SmartInviteResponse>(request);
+            return response.ToSmartInvite();
+        }
+
+        /// <inheritdoc/>
+        public SmartInviteMultiRecipient CreateInvite(SmartInviteMultiRecipientRequest smartInviteRequest)
+        {
+            var request = new HttpRequest
+            {
+                Method = "POST",
+                Url = this.urlProvider.SmartInviteUrl
+            };
+
+            request.AddOAuthAuthorization(this.clientSecret);
+            request.SetJsonBody(smartInviteRequest);
+
+            var response = this.HttpClient.GetJsonResponse<SmartInviteMultiRecipientResponse>(request);
+            return response.ToSmartInvite();
+        }
+
+        /// <inheritdoc/>
+        public SmartInviteMultiRecipient GetSmartInvite(string smartInviteId)
+        {
+            Preconditions.NotEmpty("smartInviteId", smartInviteId);
+
+            var request = new HttpRequest
+            {
+                Method = "GET",
+                Url = this.urlProvider.SmartInviteUrl,
+                QueryString = new HttpRequest.QueryStringCollection
+                {
+                    { "smart_invite_id", smartInviteId }
+                }
+            };
+
+            request.AddOAuthAuthorization(this.clientSecret);
+
+            var response = this.HttpClient.GetJsonResponse<SmartInviteMultiRecipientResponse>(request);
             return response.ToSmartInvite();
         }