+++ src/Cronofy/Requests/SmartInviteEventRequest.cs
@@ -1,7 +1,7 @@
 ﻿namespace Cronofy.Requests
 {
     /// <summary>
-    /// Class to represent a smart invite event.
+    /// Class to represent a Smart Invite event.
     /// </summary>
     public sealed class SmartInviteEventRequest : BaseEventRequest
     {