+++ src/Cronofy/Responses/SmartInviteResponse.cs
@@ -1,10 +1,11 @@
 ﻿namespace Cronofy.Responses
 {
+    using System.Collections.Generic;
     using System.Linq;
     using Newtonsoft.Json;
 
     /// <summary>
-    /// Class to represent a smart invite response.
+    /// Class to represent a Smart Invite response.
     /// </summary>
     internal sealed class SmartInviteResponse
     {
@@ -18,9 +19,9 @@
         public string Method { get; set; }
 
         /// <summary>
-        /// Gets or sets the smart invite identifier.
+        /// Gets or sets the Smart Invite identifier.
         /// </summary>
-        /// <value>The smart invite identifier.</value>
+        /// <value>The Smart Invite identifier.</value>
         [JsonProperty("smart_invite_id")]
         public string SmartInviteId { get; set; }
 
@@ -79,7 +80,11 @@
                 invite.Replies = Enumerable.Empty<SmartInvite.Attendee>();
             }
 
-            invite.Recipient = this.Recipient.ToAttendee();
+            if (this.Recipient != null)
+            {
+                invite.Recipient = this.Recipient.ToAttendee();
+            }
+
             invite.Event = this.Event.ToEvent();
             invite.Attachments = this.Attachments.ToAttachments();